Moved LoadData, Data, Download, Strategies and Plots to new module helpers.
